Explicitly set the attrs field of the Style struct to be an int64
This fixes test failures on 32bit systems where an int is only 32 bits long

+++ get.go
@@ -423,7 +423,7 @@ func (s Style) getAsBool(k propKey, defaultVal bool) bool {
 	if !s.isSet(k) {
 		return defaultVal
 	}
-	return s.attrs&int(k) != 0
+	return s.attrs&int64(k) != 0
 }
 
 func (s Style) getAsColor(k propKey) TerminalColor {
